return usize from BlockType::len_{params,results}

+++ crates/wasmi/src/engine/block_type.rs
@@ -63,22 +63,22 @@ impl BlockType {
     }
 
     /// Returns the number of parameters of the [`BlockType`].
-    pub fn len_params(&self, engine: &Engine) -> u32 {
+    pub fn len_params(&self, engine: &Engine) -> usize {
         match &self.inner {
             BlockTypeInner::Empty | BlockTypeInner::Returns => 0,
             BlockTypeInner::FuncType(func_type) => {
-                engine.resolve_func_type(func_type, |func_type| func_type.params().len() as u32)
+                engine.resolve_func_type(func_type, |func_type| func_type.params().len())
             }
         }
     }
 
     /// Returns the number of results of the [`BlockType`].
-    pub fn len_results(&self, engine: &Engine) -> u32 {
+    pub fn len_results(&self, engine: &Engine) -> usize {
         match &self.inner {
             BlockTypeInner::Empty => 0,
             BlockTypeInner::Returns => 1,
             BlockTypeInner::FuncType(func_type) => {
-                engine.resolve_func_type(func_type, |func_type| func_type.results().len() as u32)
+                engine.resolve_func_type(func_type, |func_type| func_type.results().len())
             }
         }
     }
